Customers provide billing address before payment

+++ app/controllers/checkout_controller.rb
@@ -3,11 +3,25 @@ class CheckoutController < ApplicationController
   before_action :set_locale
   before_action :set_cart, except: [:shopper_redirect] # Don't set cart in the session if we're redirecting to a sales order
   before_action :authenticate_user!, only: [:new, :shopper_redirect, :success]
+  before_action :set_address, only: [:new, :save_address]
 
   # Before checking out, let's signin/create our user so we can
   # assign the order to them after the Adyen dropin process is
   # complete
   def new
+  end
+
+  def save_address
+    respond_to do |format|
+      if @address.update(address_params)
+        format.html { redirect_to checkout_payment_path }
+      else
+        format.html { render action: "new", alert: "Please check your billing address for errors." }
+      end
+    end
+  end
+
+  def payment
   end
 
   # JS
@@ -47,11 +61,14 @@ class CheckoutController < ApplicationController
   # The payment went through, create the sales order and redirect the user there.
   # TODO: verify the shopping_cart response_data before creating the sales order
   def success
-    @sales_order = SalesOrder.create(user: current_user, shopping_cart: @shopping_cart)
-    session.delete(:shopping_cart_id)
-
+    @sales_order = SalesOrder.new(user: current_user, shopping_cart: @shopping_cart, address: current_user.addresses.last)
     respond_to do |format|
-      format.html { redirect_to @sales_order, notice: "Success! Thank you for your order. See the details below." }
+      if @sales_order.save
+        session.delete(:shopping_cart_id)
+        format.html { redirect_to @sales_order, notice: "Success! Thank you for your order. See the details below." }
+      else
+        format.html { redirect_to profile_path, alert: "We're sorry, there was a problem with your order. Please contact us using our support page." }
+      end
     end
   end
 
@@ -60,11 +77,25 @@ class CheckoutController < ApplicationController
   end
 
   def failed
-    redirect_to checkout_path, alert: "The order failed. Please try a different method of payment."
+    redirect_to checkout_payment_path, alert: "The order failed. Please try a different method of payment."
   end
 
   def error
-    redirect_to checkout_path, alert: "There was an error. Please try a different method of payment."
+    redirect_to checkout_payment_path, alert: "There was an error. Please try a different method of payment."
   end
 
+  private
+
+  def set_address
+    @address = current_user.default_address
+    @address.addressable_type = "User"
+    @address.addressable_id = current_user.id
+  end
+
+  def address_params
+    params.require(:address).permit(
+      :street_1, :street_2, :street_3, :street_4,
+      :locality, :region, :postal_code, :country
+    )
+  end
 end
